requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
contextlib2==0.5.5
coreapi==2.3.1
coreschema==0.0.4
crcmod==1.7
Django==1.10.7
django-colorfield==0.1.14
django-debug-toolbar==1.8
django-filter==1.0.4
django-solo==1.1.2
djangorestframework==3.6.3
dropbox==8.0.0
gunicorn==19.7.1
humanize==0.5.1
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
MarkupSafe==1.0
paho-mqtt==1.2.3
pkg-resources==0.0.0
psycopg2==2.7.3.1
pyserial==3.3
python-dateutil==2.6.0
pytz==2017.2
raven==6.1.0
requests==2.18.1
six==1.10.0
sqlparse==0.2.3
uritemplate==3.0.0
urllib3==1.21.1
appdirs==1.4.2
packaging==16.8
pyparsing==2.1.10
six==1.10.0
